+++ tests/test-unix-domain-to-nmsg-udp-sock-pres.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -x
+#!/bin/sh
 
 # This test reads dnstap input from a Unix domain socket
 # and writes nmsg to a UDP socket.
@@ -15,31 +15,37 @@ FSTRM_REPLAY=${FSTRM_REPLAY:-fstrm_replay}
 NMSG_DNSTAP=${NMSG_DNSTAP:-${abs_top_builddir}/src/nmsg-dnstap}
 
 # due to filename length limitation use /tmp instead of ${abs_top_builddir}
-DNSTAP_UNIX_SOCK="/tmp/${TEST}.unix.sock"
-NMSG_UDP_SOCKET="127.0.0.1/9999"
+NMSG_DNSTAP_UNIX_SOCK="/tmp/${TEST}.unix.sock"
+NMSGTOOL_UDP_SOCKET="127.0.0.1/9999"
 DNSTAP_INPUT=${abs_top_srcdir}/tests/dnstap-input-data
 
 # remove old outputs
 rm -f ${abs_top_builddir}/tests/${TEST}*.out
 
 echo start nmsgtool to listen for the nmsg via UDP
-$NMSGTOOL --readsock ${NMSG_UDP_SOCKET} > ${abs_top_builddir}/tests/${TEST}-nmsgtool.dnstap.pres.out &
+$NMSGTOOL -ddd --unbuffered --readsock ${NMSGTOOL_UDP_SOCKET} --writepres ${abs_top_builddir}/tests/${TEST}-nmsgtool.dnstap.pres.out 2>${abs_top_builddir}/tests/${TEST}-nmsgtool.dnstap.pres.stderr.out &
 NMSGTOOL_PID=$!
+
+sleep 0.25
+if ! kill -0 $NMSGTOOL_PID 2>/dev/null; then
+  echo $NMSGTOOL IS NOT RUNNING
+  exit 1
+fi
 
 # this uses default content-type
 
 echo start nmsg-dnstap to read from a Unix domain socket and write nmsg to UDP
-$NMSG_DNSTAP -ddddd --writesock ${NMSG_UDP_SOCKET} --unix ${DNSTAP_UNIX_SOCK} 2>${abs_top_builddir}/tests/${TEST}-nmsg-dnstap.stderr.out &
+$NMSG_DNSTAP -ddddd --unbuffered --writesock ${NMSGTOOL_UDP_SOCKET} --unix ${NMSG_DNSTAP_UNIX_SOCK} 2>${abs_top_builddir}/tests/${TEST}-nmsg-dnstap.stderr.out &
 NMSG_DNSTAP_PID=$!
 
 # make sure the listener is up
-sleep 0.5
+sleep 0.25
 
 echo send existing testing dnstap data using fstrm_replay to the Unix domain socket
-${FSTRM_REPLAY} --unix ${DNSTAP_UNIX_SOCK} --read-file ${DNSTAP_INPUT} --type protobuf:dnstap.Dnstap
+${FSTRM_REPLAY} --unix ${NMSG_DNSTAP_UNIX_SOCK} --read-file ${DNSTAP_INPUT} --type protobuf:dnstap.Dnstap
 
 # make sure the output is saved
-sleep 0.5
+sleep 0.25
 
 # stop the processes
 kill $NMSG_DNSTAP_PID
@@ -61,8 +67,6 @@ else
   status=1
   echo FAIL
 fi
-
-diff -u ${abs_top_srcdir}/tests/nmsgtool.dnstap.pres ${abs_top_builddir}/tests/${TEST}-nmsgtool.dnstap.pres.no-timestamps.out
 
 exit $status
